Made routes for state change and data posting
Created functions that execute every second to check the state of
the game in the database. Created routes to post data about plays and
games. 2 players can now join the same lobby, and play a game turn for
turn. The state updates in the other browser shortly after the play is
made on the active browser.