Handled player move posting and game reset
Players can now alternate turns and log their plays in the database.
When a winner is found, the winner_id of the game record is changed to
the id of the game winner.

Only the game winner can start a new game, and the loser waits using a
callback function on a timer. When the new game starts, player tokens
are swapped and the player with the token 'X' goes first.